Reorganize keyset classes
