+++ spec/lib/keyset/desc_spec.rb
@@ -2,7 +2,7 @@
 
 require "spec_helper"
 
-describe GraphQL::Connections::KeysetAsc do
+describe GraphQL::Connections::Keyset::Desc do
   let_it_be(:msg_a) { Message.create!(body: "A", created_at: Time.local(2020, 10, 2, 15)) }
   let_it_be(:msg_b) { Message.create!(body: "B", created_at: Time.local(2020, 10, 2, 14)) }
   let_it_be(:msg_c) { Message.create!(body: "C", created_at: Time.local(2020, 10, 2, 13)) }
@@ -22,7 +22,7 @@ describe GraphQL::Connections::KeysetAsc do
 
     it "returns default_max_page_size nodes" do
       expect(nodes.size).to eq 3
-      expect(names).to eq %w[A B C]
+      expect(names).to eq %w[E D C]
       expect(connection.has_previous_page).to be false
       expect(connection.has_next_page).to be true
     end
@@ -33,7 +33,7 @@ describe GraphQL::Connections::KeysetAsc do
 
     it "returns first two nodes" do
       expect(nodes.size).to eq 2
-      expect(names).to eq %w[A B]
+      expect(names).to eq %w[E D]
       expect(connection.has_previous_page).to be false
       expect(connection.has_next_page).to be true
     end
@@ -44,104 +44,104 @@ describe GraphQL::Connections::KeysetAsc do
 
     it "returns last two nodes" do
       expect(nodes.size).to eq 2
-      expect(names).to eq %w[D E]
+      expect(names).to eq %w[B A]
       expect(connection.has_previous_page).to be true
       expect(connection.has_next_page).to be false
     end
   end
 
   describe ":after param" do
-    let(:params) { {after: base_connection.cursor_for(msg_b)} }
+    let(:params) { {after: base_connection.cursor_for(msg_d)} }
 
     it "returns default_max_page_size nodes after B" do
       expect(nodes.size).to eq 3
-      expect(names).to eq %w[C D E]
+      expect(names).to eq %w[C B A]
       expect(connection.has_previous_page).to be true
       expect(connection.has_next_page).to be false
     end
   end
 
   describe ":before param" do
-    let(:params) { {before: base_connection.cursor_for(msg_d)} }
+    let(:params) { {before: base_connection.cursor_for(msg_b)} }
 
     it "returns default_max_page_size nodes before D" do
       expect(nodes.size).to eq 3
-      expect(names).to eq %w[A B C]
+      expect(names).to eq %w[E D C]
       expect(connection.has_previous_page).to be false
       expect(connection.has_next_page).to be true
     end
   end
 
   describe ":after and :first params" do
-    let(:params) { {after: base_connection.cursor_for(msg_b), first: 2} }
+    let(:params) { {after: base_connection.cursor_for(msg_d), first: 2} }
 
     it "returns two nodes after B" do
       expect(nodes.size).to eq 2
-      expect(names).to eq %w[C D]
+      expect(names).to eq %w[C B]
       expect(connection.has_previous_page).to be true
       expect(connection.has_next_page).to be true
     end
   end
 
   describe ":before and :last params" do
-    let(:params) { {before: base_connection.cursor_for(msg_d), last: 2} }
+    let(:params) { {before: base_connection.cursor_for(msg_b), last: 2} }
 
-    it "returns two nodes before D" do
+    it "returns two nodes before B" do
       expect(nodes.size).to eq 2
-      expect(names).to eq %w[B C]
+      expect(names).to eq %w[D C]
       expect(connection.has_previous_page).to be true
       expect(connection.has_next_page).to be true
     end
   end
 
   describe ":after and :last params" do
-    let(:params) { {after: base_connection.cursor_for(msg_b), last: 2} }
+    let(:params) { {after: base_connection.cursor_for(msg_d), last: 2} }
 
-    it "returns two last nodes after B" do
+    it "returns two last nodes after D" do
       expect(nodes.size).to eq 2
-      expect(names).to eq %w[D E]
+      expect(names).to eq %w[B A]
       expect(connection.has_previous_page).to be true
       expect(connection.has_next_page).to be false
     end
   end
 
   describe ":before and :first params" do
-    let(:params) { {before: base_connection.cursor_for(msg_d), first: 2} }
+    let(:params) { {before: base_connection.cursor_for(msg_b), first: 2} }
 
-    it "returns two first nodes before D" do
+    it "returns two first nodes before B" do
       expect(nodes.size).to eq 2
-      expect(names).to eq %w[A B]
+      expect(names).to eq %w[E D]
       expect(connection.has_previous_page).to be false
       expect(connection.has_next_page).to be true
     end
   end
 
   describe ":after, :first, and :last params" do
-    let(:params) { {after: base_connection.cursor_for(msg_b), first: 2, last: 2} }
+    let(:params) { {after: base_connection.cursor_for(msg_d), first: 2, last: 2} }
 
-    it "returns three nodes after B" do
+    it "returns three nodes after D" do
       expect(nodes.size).to eq 3
-      expect(names).to eq %w[C D E]
+      expect(names).to eq %w[C B A]
       expect(connection.has_previous_page).to be true
       expect(connection.has_next_page).to be false
     end
   end
 
   describe ":before, :last, and :first params" do
-    let(:params) { {before: base_connection.cursor_for(msg_d), first: 2, last: 2} }
+    let(:params) { {before: base_connection.cursor_for(msg_b), first: 2, last: 2} }
 
-    it "returns three nodes before D" do
+    it "returns three nodes before B" do
       expect(nodes.size).to eq 3
-      expect(names).to eq %w[A B C]
+      expect(names).to eq %w[E D C]
       expect(connection.has_previous_page).to be false
       expect(connection.has_next_page).to be true
     end
   end
 
   describe ":after, :before, and :first params" do
-    let(:params) { {after: base_connection.cursor_for(msg_b), before: base_connection.cursor_for(msg_e), first: 1} }
+    let(:params) { {after: base_connection.cursor_for(msg_d), before: base_connection.cursor_for(msg_a), first: 1} }
 
-    it "returns first node after B and before E" do
+    it "returns first node after D and before A" do
       expect(nodes.size).to eq 1
       expect(names).to eq ["C"]
       expect(connection.has_previous_page).to be true
@@ -150,11 +150,11 @@ describe GraphQL::Connections::KeysetAsc do
   end
 
   describe ":after, :before, and :last params" do
-    let(:params) { {after: base_connection.cursor_for(msg_b), before: base_connection.cursor_for(msg_e), last: 1} }
+    let(:params) { {after: base_connection.cursor_for(msg_d), before: base_connection.cursor_for(msg_a), last: 1} }
 
-    it "returns last node after B and before E" do
+    it "returns last node after D and before A" do
       expect(nodes.size).to eq 1
-      expect(names).to eq ["D"]
+      expect(names).to eq ["B"]
       expect(connection.has_previous_page).to be true
       expect(connection.has_next_page).to be false
     end
@@ -162,12 +162,12 @@ describe GraphQL::Connections::KeysetAsc do
 
   describe ":after, :before, :first, and :last params" do
     let(:params) do
-      {after: base_connection.cursor_for(msg_b), before: base_connection.cursor_for(msg_e), first: 1, last: 1}
+      {after: base_connection.cursor_for(msg_d), before: base_connection.cursor_for(msg_a), first: 1, last: 1}
     end
 
-    it "returns first and last nodes after B and before E" do
+    it "returns first and last nodes after D and before A" do
       expect(nodes.size).to eq 2
-      expect(names).to eq %w[C D]
+      expect(names).to eq %w[C B]
       expect(connection.has_previous_page).to be true
       expect(connection.has_next_page).to be true
     end
